Вынесение скриптов под адаптив в отдельный файл
