+++ src/index.js
@@ -6,3 +6,4 @@ import './_our-projects.js';
 import './_footer-year.js';
 import './_animation.js';
 import './_policy.js';
+import './_adaptive.js';
